@@ -10,27 +10,27 @@
 	"private": false,
 	"dependencies": {
 		"lodash": "^4.17.21",
-		"qrcode": "^1.4.4",
-		"react": "^17.0.2",
-		"react-dom": "^17.0.2",
-		"react-feather": "^2.0.9",
-		"regenerator-runtime": "^0.13.9"
+		"qrcode": "^1.5.3",
+		"react": "^18.2.0",
+		"react-dom": "^18.2.0",
+		"react-feather": "^2.0.10",
+		"regenerator-runtime": "^0.14.0"
 	},
 	"devDependencies": {
-		"@svgr/webpack": "^5.5.0",
-		"@swc/cli": "^0.1.51",
-		"@swc/core": "^1.2.93",
-		"css-loader": "^6.3.0",
+		"@svgr/webpack": "^8.1.0",
+		"@swc/cli": "^0.1.62",
+		"@swc/core": "^1.3.83",
+		"css-loader": "^6.8.1",
 		"file-loader": "^6.2.0",
-		"html-webpack-plugin": "^5.3.2",
-		"mini-css-extract-plugin": "^2.4.2",
-		"sass": "^1.42.1",
-		"sass-loader": "^12.1.0",
-		"style-loader": "^3.3.0",
-		"swc-loader": "^0.1.15",
-		"webpack": "^5.58.0",
-		"webpack-cli": "^4.10.0",
-		"webpack-dev-server": "^4.3.1"
+		"html-webpack-plugin": "^5.5.3",
+		"mini-css-extract-plugin": "^2.7.6",
+		"sass": "^1.66.1",
+		"sass-loader": "^13.3.2",
+		"style-loader": "^3.3.3",
+		"swc-loader": "^0.2.3",
+		"webpack": "^5.88.2",
+		"webpack-cli": "^5.1.4",
+		"webpack-dev-server": "^4.15.1"
 	},
 	"scripts": {
 		"start": "npx webpack serve --open",
